Fixed wrong numpy type in quant op

+++ src/finn/custom_op/general/quant.py
@@ -99,8 +99,8 @@ def quant(inp_tensor, scale, zeropt, bitwidth, signed, narrow):
     min_int_val = min_int(signed, narrow, bitwidth)
     max_int_val = max_int(signed, narrow, bitwidth)
     #re-impl of: y_int = self.tensor_clamp_impl(y_int, min_val=min_int_val, max_val=max_int_val)
-    y_int = np.where(y_int > max_int_val, max_int_val.type_as(y_int), y_int)
-    y_int = np.where(y_int < min_int_val, min_int_val.type_as(y_int), y_int)
+    y_int = np.where(y_int > max_int_val, max_int_val.astype(y_int.dtype), y_int)
+    y_int = np.where(y_int < min_int_val, min_int_val.astype(y_int.dtype), y_int)
     #re-impl of: y_int = self.float_to_int_impl(y_int)
     y_int = np.round(y_int)
